python-service/requirements.txt
Flask
Flask-CORS
mysql-connector-python
textblob
python-dotenv
openai